Revert "Improve startup performance by deferring load of urllib3. (#124)"
This reverts commit b709b00de6d2707807631a1edc286ff6ec7506a4.
