--- newrelic/common/constants.py
@@ -1,9 +1,0 @@
-LOCALHOST_EQUIVALENTS = frozenset((
-    'localhost',
-    '127.0.0.1',
-    '0.0.0.0',
-    '0:0:0:0:0:0:0:0',
-    '0:0:0:0:0:0:0:1',
-    '::1',
-    '::',
-))
