+++ newrelic/core/database_node.py
@@ -17,7 +17,7 @@ from collections import namedtuple
 import newrelic.core.attribute as attribute
 import newrelic.core.trace_node
 
-from newrelic.common import constants, system_info
+from newrelic.common import system_info
 from newrelic.core.database_utils import sql_statement, explain_plan
 from newrelic.core.node_mixin import DatastoreNodeMixin
 from newrelic.core.metric import TimeMetric
@@ -67,7 +67,7 @@ class DatabaseNode(_DatabaseNode, DatastoreNodeMixin):
 
     @property
     def instance_hostname(self):
-        if self.host in constants.LOCALHOST_EQUIVALENTS:
+        if self.host in system_info.LOCALHOST_EQUIVALENTS:
             hostname = system_info.gethostname()
         else:
             hostname = self.host
