+++ tests/cross_agent/test_boot_id_utilization_data.py
@@ -18,7 +18,7 @@ import pytest
 import sys
 import tempfile
 
-from newrelic.common.utilization import BootIdUtilization
+from newrelic.common.system_info import BootIdUtilization
 
 from testing_support.fixtures import validate_internal_metrics
